manifest: mechanical move of committedManifestManager methods to new file
